1.Added Service Interfaces for Class, Studentdetails, Subject and Teacher Service classes. 2. Added Student and Subject Console UI page. 3. Changed contactnumber in StudentDetail from int to long

+++ School.Data/Models/StudentDetail.cs
@@ -14,7 +14,7 @@
 
 		public string Address { get; set; }
 
-		public int ContactNumber { get; set; }
+		public long ContactNumber { get; set; }
 
 		public StudentStatus Status { get; set; }
 	}
